Move field config to field directory
